+++ Sources/Extensions/ArrayExtension.swift
@@ -17,21 +17,7 @@ extension Array {
 
 // MARK: - Remove
 
-extension Array where Element : Equatable {
-
-    public mutating func remove(object: Iterator.Element) -> Bool {
-        if let index = self.index(of: object) {
-            self.remove(at: index)
-            return true
-        }
-        return false
-    }
-
-    public mutating func remove(objects: Element) {
-        for idx in self.indexes(of: objects).reversed() {
-            self.remove(at: idx)
-        }
-    }
+extension Array where Element: Equatable {
 
     public mutating func removeDuplicates() {
         self = reduce([]) { $0.contains($1) ? $0 : $0 + [$1] }
@@ -41,34 +27,11 @@ extension Array where Element : Equatable {
         return reduce([]) { $0.contains($1) ? $0 : $0 + [$1] }
     }
 
-    public mutating func removeAll(_ item: Element) {
-        self = filter { $0 != item }
-    }
-
-    public mutating func removedAll(_ item: Element) -> [Element] {
-        return filter { $0 != item }
-    }
-
-}
-
-// MARK: - Getter
-
-extension Array {
-
-    public func random() -> Element? {
-        guard isNotEmpty else {
-            return nil
-        }
-
-        let index = Int(arc4random_uniform(UInt32(count)))
-        return self[index]
-    }
-
 }
 
 // MARK: - Index Getter
 
-extension Array where Element : Equatable {
+extension Array where Element: Equatable {
 
     public func indexes(of item: Element) -> [Int] {
         var indexes = [Int]()
@@ -93,7 +56,7 @@ extension Array where Element : Equatable {
 
 // MARK: - Equatable Transform
 
-extension Array where Element : Equatable {
+extension Array where Element: Equatable {
 
     public func difference(with values: [Element]...) -> [Element] {
         var result = [Element]()
@@ -112,8 +75,8 @@ extension Array where Element : Equatable {
         var result = self
         var intersection = Array()
 
-        for (i, value) in values.enumerated() {
-            if i > 0 {
+        for (index, value) in values.enumerated() {
+            if index > 0 {
                 result = intersection
                 intersection = Array()
             }
@@ -144,62 +107,6 @@ extension Array where Element : Equatable {
             let endIndex = ($0.advanced(by: chunkSize) > self.count) ? self.count - $0 : chunkSize
             return Array(self[$0..<$0.advanced(by: endIndex)])
         }
-    }
-
-}
-
-// MARK: - Transform
-
-extension Array {
-
-    public func shuffled() -> [Element] {
-        var array = self
-        array.shuffle()
-        return array
-    }
-
-    public mutating func shuffle() {
-        guard count > 2 else {
-            return
-        }
-
-        for idx in startIndex..<endIndex {
-            let next = Int(arc4random_uniform(UInt32(endIndex - idx))) + idx
-            guard idx != next else {
-                continue
-            }
-            self.swapAt(idx, next)
-        }
-    }
-
-}
-
-// MARK: - Misc
-
-extension Array {
-
-    public func testAll(test: (Element) -> Bool) -> Bool {
-        for item in self {
-            if !test(item) {
-                return false
-            }
-        }
-        return true
-    }
-
-}
-
-// MARK: - Misc Equatable
-
-extension Array where Element : Equatable {
-
-    public func contains(_ elements: [Element]) -> Bool {
-        for item in elements {
-            if contains(item) == false {
-                return false
-            }
-        }
-        return true
     }
 
 }