+++ Sources/Extensions/UIKit/UILabelExtension.swift
@@ -15,10 +15,10 @@ extension UILabel {
         paragraphStyle.lineHeightMultiple = lineHeight
         paragraphStyle.alignment = textAlignment
         paragraphStyle.lineBreakMode = lineBreakMode
-    
+
         let attrString = NSMutableAttributedString(string: text!)
-        attrString.addAttribute(NSAttributedStringKey.font, value: font, range: NSRange(location: 0, length: attrString.length))
-        attrString.addAttribute(NSAttributedStringKey.paragraphStyle, value: paragraphStyle, range: NSRange(location: 0, length: attrString.length))
+        attrString.addAttribute(NSAttributedString.Key.font, value: font, range: NSRange(location: 0, length: attrString.length))
+        attrString.addAttribute(NSAttributedString.Key.paragraphStyle, value: paragraphStyle, range: NSRange(location: 0, length: attrString.length))
         attributedText = attrString
     }
 
@@ -51,7 +51,7 @@ extension UILabel {
         let rectSize = CGSize(width: self.width, height: .greatestFiniteMagnitude)
         let size: CGSize = (string as NSString).boundingRect(with: rectSize,
                                                              options: .usesLineFragmentOrigin,
-                                                             attributes: [NSAttributedStringKey.font: font],
+                                                             attributes: [NSAttributedString.Key.font: font],
                                                              context: nil).size
         return (size.height > self.bounds.size.height)
     }